requirements.txt
cos_python_sdk_v5==1.9.25
fastapi~=0.100.0
openai==0.27.8
Pillow==10.0.0
Pillow==10.0.0
pydantic~=2.0.2
SQLAlchemy==2.0.19
tweepy==4.14.0
webuiapi==0.9.5

redis~=4.6.0
celery~=5.3.1